WIP/MINOR: thread: always reorder CPUs
Previously CPUs were only reordered when absolutely no option was set.
It makes more sense to do it all the time otherwise we cannot for
example use adjacent CPUs when the number of threads is limited.

WIP: The thread filling algo should be redone completely.

The principle is to start from a pool of sorted CPUs on one side, and
a pool of min..max thread groups on the other side, and distribute CPUs
up to thread.max between the groups based on affinity and topology.
Thread groups can be added on the fly up to maxgrp if more clusters
are discovered or more than 64 threads appear in a group. Then, up to
thread.max can be added if some interesting CPUs are discovered and
there's room left. Then if thread.min was not reached, the best
remaining CPUs should be added to existing groups based on topology.