WIP/MINOR: cpuset: add a CPU selection strategy setting to the global section
We'll need to let the user decide what's best for their workload, and in
order to do this we'll have to provide tunable options. For that, we're
introducing struct ha_cpu_selection which contains a name, a description
and a function pointer. The purpose will be to use that function pointer
to sort the threads before limiting them to the desired number. Right
now, only "performance" and "all" are supported. "performance" enumerates
the performant cores before the other ones. "all" keeps them in their
original order.

The CPUs are sorted according to the configured criterion before being
listed.