MEDIUM: numa: eliminate undesired CPUs from the listed ones
Now that we've sorted eligible CPUs in preference order, we can drop
the extraneous ones. At the moment the result is bogus if we let the
numa mapping perform the counting because it only counts those on the
same node without any consideration for the ordering, so that can give
for example 28 (randomly placed) without considering the users' choice.
This code should now only be conditioned by the CPU selection strategy
and not run in all cases. But it allows to select the best subset of
CPUs for a given thread count.