WIP/MINOR: numa: assign different cluster IDs to different L3
It's important to report a different cluster ID to different L3s.
We could also have used a different entry to set these IDs but on
some CPUs these clusters report some info (e.g. E-cores on 14900)
but on others it incorrectly reports one core (e.g. P-cores on the
same SoC).

This is currently done in the linux-specific code but it's not the
correct place, it should be done in a generic place (except if we
consider that we're fixing linux's report but that's not really
the case).